requirements.txt
groq
python-dotenv
supabase